+++ src/main/java/com/fptu/paa/constant/RoleName.java
@@ -3,5 +3,5 @@ package com.fptu.paa.constant;
 public enum RoleName {
 	ROLE_CUSTOMER,
 	ROLE_OWNER,
-	ROLE_ADMIN,
+	ROLE_STAFF,
 }
